MouseImplementation/requirements.txt
PyDirectInput==1.0.4
pyserial==3.5
